"ads1_m14_a1: 0/4 testcases passed. Cannot run hidden testcases as Server is unreachablecheck_style score: 0/1 "

+++ ADS-1-assignments/m14/14.1 Dynamic Median Finding/Solution.java
@@ -27,9 +27,9 @@ final class Solution {
 			float median = ((float)list[(i - 1)/2] + (float)list[(i/2)/2]);
 				System.out.println(median);
 		} else if(i == 2) {
-			System.out.println((list[0] + list[1]) / 2);
+			System.out.println((float)(list[0] + list[1]) / 2);
 		} else {
-			System.out.println(list[0]);
+			System.out.println((float)list[0]);
 		}
 		n--;
 		}
